// src/types/TSteamSpy.ts
type SteamSpyDataRes = {
    appid: number;
    name: string;
    developer: string;
    publisher: string;
    score_rank: string;
    positive: number;
    negative: number;
    userscore: number;
    owners: string;
    average_forever: number;
    average_2weeks: number;
    median_forever: number;
    median_2weeks: number;
    price: string;
    initialprice: string;
    discount: string;
    ccu: number;
    languages: string;
    genre: string[];
    tags: {
        [key: string]: number;
    };
}[];

export default SteamSpyDataRes;